adds COI sample graph
